+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ##v1.2.0 - unreleased
  - Added NeedlessNullnessDetector to report on needless @Nullable / @NonNull
     annotations on primitive / void parameters and return types.
+ - Added `ViewInflateDetector` to report issues on calls to `View.inflate`
+    prior to Lollipop
 
 ##v1.1.9
  - Fix compatibility with latest android tools.